Squashed commit of the following:
commit 76ec6f75cbae064f4ca2658c45411b9a6647ab57
Author: gary <gary.ng@robocoder.com>
Date:   Thu Apr 8 19:52:12 2021 -0700

    R11.242.10408 community version installer

commit 4ed78cd267d2fdd53bb7c635726c27044d4c3a82
Author: gary <gary.ng@robocoder.com>
Date:   Thu Apr 8 19:43:03 2021 -0700

    Squashed commit of the following:

    commit a49321ace7c478622cf314f044d732b5d81efcad
    Author: gary <gary.ng@robocoder.com>
    Date:   Wed Mar 24 10:15:40 2021 -0700

        dynamic page title fix

    commit e552d6a63e3b7bc476e44472598e643f43fef3aa
    Author: gary <gary.ng@robocoder.com>
    Date:   Tue Mar 23 00:47:56 2021 -0700

        unique index can only be dropped as constraint not index

    commit 1ddec0ebbe2054001c0f406f6d53aeebbc0755d9
    Author: gary <gary.ng@robocoder.com>
    Date:   Wed Mar 17 07:17:16 2021 -0700

        new company preference screen

    commit 8e1184480eefed927caae9a6d3263fba92c8fe92
    Author: gary <gary.ng@robocoder.com>
    Date:   Wed Mar 10 22:27:34 2021 -0800

        enhance deployment to handle required IUSR read for web layer and better password encryption handling in new installation

    commit 2dcf3d40a973e076a8ae79bb70896b31ec95c678
    Author: gary <gary.ng@robocoder.com>
    Date:   Thu Mar 4 03:57:50 2021 -0800

        further enhancement for deploy/install to handle unconventional use of 'create VIEW' or roDesign.dbo. like construct.

    commit bf2252b75edfc6c6118bb3f4a00eca5133b1f51b
    Author: aaron xu <aaron.xu@robocoder.com>
    Date:   Wed Mar 3 15:10:16 2021 -0800

        1. flowchart IDE
           - add server rule (Ir_UpdFlowchartDesc)
           - remove system list
           - add company filter
        2. viewflowchart screen modification

    commit a48f25cf026405767cb859d16167744b7f532b09
    Author: gary <gary.ng@robocoder.com>
    Date:   Tue Mar 2 20:11:08 2021 -0800

        remove dependency of Rule3 on WebRules

    commit ac04bbdedaa421f038c2bb8eb15242179a8b0177
    Author: gary <gary.ng@robocoder.com>
    Date:   Tue Mar 2 19:51:13 2021 -0800

        restructure WebAuthn feature to remove unnecessary assembly dependency in WebRules(causing silent not copying dll in command line msbuild)

    commit 5e01c926c45ad1befe96da2699b0b08baa502a38
    Author: gary <gary.ng@robocoder.com>
    Date:   Tue Mar 2 19:49:31 2021 -0800

        restructure project to remove unnecessary inter project dependency from Rule3 to WebRules

    commit 7fc4d0930ef8e431188986ca850ec12a8257c3a5
    Author: gary <gary.ng@robocoder.com>
    Date:   Tue Mar 2 17:37:47 2021 -0800

        add CrOneCompDefaultOnly to ROCmonD for deployment to all modules metadata db

    commit 2076e205a0308b5a631292e417ecfb7e54ced1c8
    Author: gary <gary.ng@robocoder.com>
    Date:   Tue Mar 2 17:27:39 2021 -0800

        force full rebuild on git checkout via admlicense webhook

    commit 1928d791d307c26d5f2e06fdd7b395c6fe7f29d8
    Author: gary <gary.ng@robocoder.com>
    Date:   Tue Mar 2 05:30:38 2021 -0800

        add helper function to convert NewtonSoft Dynamic JContainer tree to C# object tree of List/Dictionary(for serialization)

    commit 6d450fbec01e15294d3dbb799b121c2a863df40e
    Author: aaron xu <aaron.xu@robocoder.com>
    Date:   Mon Mar 1 15:18:58 2021 -0800

        Handle null exception for the LCurr.CompanyId.ToString() when sign out

    commit babc7f302ae52cd27f5acf01d2ab2abb7ec82d35
    Author: gary <gary.ng@robocoder.com>
    Date:   Fri Feb 26 18:40:42 2021 -0800

        handle exception case of not able to read company profile

    commit 0093019fae12c5778e09ee35998b60b900dd4a22
    Author: aaron xu <aaron.xu@robocoder.com>
    Date:   Thu Feb 25 15:30:27 2021 -0800

        1. header generation code (handle the case if there is no image uploaded for the company)
        2. Modify the flowchart webrules
        3. Add server rule (CrOneCompPrefOnly) for company preference screen and modify the CrOneCompDefaultOnly
        4. Modify the view flowchart module (handle the case when no company selected)

    commit 03cdeb10a4983f8eba75e420a3ef18be67f10ded
    Author: gary <gary.ng@robocoder.com>
    Date:   Wed Feb 24 17:39:10 2021 -0800

        only load company profile once per session

    commit 1e1650f79cff4efffc375e7d09918abb7688ea32
    Author: gary <gary.ng@robocoder.com>
    Date:   Wed Feb 24 16:11:47 2021 -0800

        add company profile mechanism

    commit d890f711fd72517e0f654c0a1a31dcb53f8d9aba
    Author: aaron xu <aaron.xu@robocoder.com>
    Date:   Wed Feb 24 12:39:15 2021 -0800

        flowchart modification, new company preference screen, default and header modules modifications

    commit b3536f72959a6e64243af66c82c473b16ee317a6
    Author: gary <gary.ng@robocoder.com>
    Date:   Mon Feb 22 22:00:29 2021 -0800

        enhance installer backup with more progress info for tracking, version 4.9.19.0

    commit 1f2751c5cb83bf6d4219a78c809c62a255746ee2
    Author: aaron xu <aaron.xu@robocoder.com>
    Date:   Mon Feb 22 12:51:06 2021 -0800

        Two new columns added to flowchart table: companyId, companyDefault

+++ Rule3/Dbscript.cs
@@ -505,9 +505,21 @@ ORDER BY so1.name"
                         }
                         hasContent = true;
                         sbDrop.Append("IF EXISTS (SELECT i.name FROM sysindexes i INNER JOIN sysobjects o ON i.id = o.id WHERE i.name = '" + drIx[0].ToString() + "' AND o.name = '" + dr2["tbName"].ToString() + "')\r\n");
-						sbDrop.Append("    DROP INDEX " + dr2["tbName"].ToString() + "." + drIx[0].ToString() + " \r\n\r\n");
+                        if (drIx[1].ToString().LastIndexOf("unique") > 0 && 1 != 1)
+                        {
+                            // some unique key was created as constraint and not index which should be drop as contstrain
+                            // though no way to tell from development thus disabled
+                            // can only be done manually and re-run to re-create as unique index !!!
+                            sbDrop.Append("    ALTER TABLE dbo." + dr2["tbName"].ToString() + " DROP CONSTRAINT " + drIx[0].ToString() + " \r\n\r\n");
+                        }
+                        else
+                        {
+                            sbDrop.Append("    DROP INDEX " + dr2["tbName"].ToString() + "." + drIx[0].ToString() + " \r\n\r\n");
+                        }
 						strIx = "CREATE ";
 						if (drIx[1].ToString().LastIndexOf("unique") > 0) {strIx += " UNIQUE ";}
+                        // there are case where non-primary index is clustered !
+                        if (drIx[1].ToString().LastIndexOf("nonclustered") < 0) { strIx += " CLUSTERED "; }
 						strIx += "INDEX " + drIx[0].ToString() + " ON " + dr2["tbName"].ToString() + "(";
 						strIx += drIx[2].ToString() + ")" 
                                     /* add covering columns */
@@ -594,7 +606,7 @@ ORDER BY so1.name"
 				foreach (DataRow dr2 in dtVw.Rows) {sbView.Append(dr2[0].ToString());}
 				if (!sbView.ToString().Equals(""))
 				{
-                    Regex rx = new Regex("(CREATE VIEW)(\\s+.*)((\\[)?" + dr[0].ToString() + "(\\])?)(.*\\s+AS)");
+                    Regex rx = new Regex("(CREATE VIEW)(\\s+.*)((\\[)?" + dr[0].ToString() + "(\\])?)(.*\\s+AS)",RegexOptions.IgnoreCase);
                     //sbCrea.Append("if exists (select * from dbo.sysobjects where id = object_id(N'dbo." + dr[0].ToString() + "') and OBJECTPROPERTY(id, N'IsView') = 1)\r\n");
                     //sbCrea.Append("drop view dbo." + dr[0].ToString() + "\r\n");
                     //sbCrea.Append("GO\r\n");
@@ -629,6 +641,7 @@ ORDER BY so1.name"
 				{
                     // we use the [name] form to distinguish between hand coded string from sp_helptext
                     Regex rx = new Regex("(CREATE\\s+PROCEDURE)(\\s+[^+]*)((\\[)?" + dr[0].ToString() + "(\\])?)", RegexOptions.Multiline | RegexOptions.IgnoreCase);
+                    Regex rxView = new Regex("(CREATE\\s+VIEW)(\\s+[^+]*)((\\[)?" + dr[0].ToString() + "(\\])?)", RegexOptions.Multiline | RegexOptions.IgnoreCase);
                     sb.Append(ScriptDropSp(dr[0].ToString(), dr[1].ToString().Trim()));
 					sb.Append("GO\r\n");
 					sb.Append("SET QUOTED_IDENTIFIER ON\r\n");
@@ -636,6 +649,14 @@ ORDER BY so1.name"
 					if (SrcDbProviderCd == "S") {sb.Append("SET ANSINULL ON\r\n");} else {sb.Append("SET ANSI_NULLS ON\r\n");}
 					sb.Append("GO\r\n");
                     ss = ss.Trim(new char[] { ' ', '\r', '\n' }).Replace("\r\n","\r").Replace("\n","\r").Replace("\r",Environment.NewLine);
+                    ss = rx.Replace(ss, (m) =>
+                    {
+                        return "ALTER PROCEDURE" + m.Groups[2].Value + m.Groups[3].Value;
+                    });
+                    ss = rx.Replace(ss, (m) =>
+                    {
+                        return "ALTER FUNCTION" + m.Groups[2].Value + m.Groups[3].Value;
+                    });
                     ss = rx.Replace(ss, (m) =>
                     {
                         return "ALTER PROCEDURE" + m.Groups[2].Value + m.Groups[3].Value;
